Updated the README. Working on formatting.
ChordBuilder is now a complete widget. It emits a signal called submit
which contains a string representing the chord to be submitted. I have
temporarily commented out some of the buttons that are not compatible
with the chord parsing regex.

Started working on UserProgress. It now contains a grid of all known
chords and the ChordBuilder widget. You can add chords to the list using
the chord builder and the chord grid will update. Still in testing and
does not save the data.

Changed one comment in chordchanges.py

Added a todo in chorddata.py. I want the add_chord method to specify why
a chord was not added.